Move pluralizer and url generator to api package
These are utility classes that rather belong to a util package,
but a dedicated package for that doesn't seem worth it.
