+++ packages/kubernetes-client/src/api/pluralizer.spec.ts
