+++ packages/kubernetes-client/src/api/pluralizer.ts
