+++ packages/kubernetes-client/src/fetch/index.ts
@@ -1,6 +1,4 @@
 export * from './client'
 export * from './builder'
 export * from './client'
-export * from './urlgenerator'
 export * from './authorizer'
-export * from './pluralizer'
